Refactor : Ladder 클래스 변경
불필요한 클래스를 삭제, 기존의 LadderMaker 클래스를 Ladder 클래스로 변경
