--- src/main/java/kr/codesquad/domain/laddergenerator/LadderMaker.java
@@ -1,34 +1,0 @@
-package kr.codesquad.domain.laddergenerator;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
-
-public class LadderMaker {
-
-    private List<Line> ladder;
-    private int people;
-    private int height;
-
-
-    public String makeLadder(List<String> people, int height){
-        this.people = people.size() - 1;
-        this.height = height;
-
-        ladder = new ArrayList<>();
-        for (int i = 0; i < height; i++) {
-            ladder.add(new Line(this.people));
-        }
-        return makeLadderToString();
-    }
-
-    private String makeLadderToString(){
-        StringBuilder ladderResult = new StringBuilder();
-
-        for (int i = 0; i < ladder.size(); i++) {
-            ladderResult.append(ladder.get(i).makeLineShape()).append("\n");
-        }
-        return ladderResult.toString();
-    }
-
-}
